+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - Fix inline-def to work with defs with metas.
 - Bump clj-kondo to `2026.04.15`.
 - bump up timeout for code action performance measurement, include p90 measurement #2236
+- implementation of inline function. #1827
 - Fix initialization crash when a source file has syntax errors (e.g. unbalanced parens) by using safe parser in unused-public-var linter's `:gen-class` check. #2242
 - Bump rewrite-clj to `1.2.54`.
 - implement move to :let refactoring #1732